+++ R/utilities.R
@@ -1339,18 +1339,11 @@ PseudobulkExpression <- function(
     if (slot[1] == 'scale.data') {
       na.matrix <- data.return[[1]]
       na.matrix[1:length(x = na.matrix)] <- NA
-      # TODO: restore once check.matrix is in SeuratObject
-      # toRet <- CreateSeuratObject(
-      #   counts = na.matrix,
-      #   project = if (pb.method == "average") "Average" else "Aggregate",
-      #   assay = names(x = data.return)[1],
-      #   check.matrix = FALSE,
-      #   ...
-      # )
       toRet <- CreateSeuratObject(
         counts = na.matrix,
         project = if (pb.method == "average") "Average" else "Aggregate",
         assay = names(x = data.return)[1],
+        check.matrix = FALSE,
         ...
       )
       toRet <- SetAssayData(
@@ -1372,18 +1365,11 @@ PseudobulkExpression <- function(
         new.data = data.return[[1]]
       )
     } else {
-      # TODO: restore once check.matrix is in SeuratObject
-      # toRet <- CreateSeuratObject(
-      #   counts = data.return[[1]],
-      #   project = if (pb.method == "average") "Average" else "Aggregate",
-      #   assay = names(x = data.return)[1],
-      #   check.matrix = FALSE,
-      #   ...
-      # )
       toRet <- CreateSeuratObject(
         counts = data.return[[1]],
         project = if (pb.method == "average") "Average" else "Aggregate",
         assay = names(x = data.return)[1],
+        check.matrix = FALSE,
         ...
       )
       toRet <- SetAssayData(
@@ -1399,9 +1385,7 @@ PseudobulkExpression <- function(
         if (slot[i] == 'scale.data') {
           na.matrix <- data.return[[i]]
           na.matrix[1:length(x = na.matrix)] <- NA
-          # TODO: restore once check.matrix is in SeuratObject
-          # toRet[[names(x = data.return)[i]]] <- CreateAssayObject(counts = na.matrix, check.matrix = FALSE)
-          toRet[[names(x = data.return)[i]]] <- CreateAssayObject(counts = na.matrix)
+          toRet[[names(x = data.return)[i]]] <- CreateAssayObject(counts = na.matrix, check.matrix = FALSE)
           toRet <- SetAssayData(
             object = toRet,
             assay = names(x = data.return)[i],
@@ -1421,9 +1405,7 @@ PseudobulkExpression <- function(
             new.data = as.matrix(x = data.return[[i]])
           )
         } else {
-          # TODO: restore once check.matrix is in SeuratObject
-          # toRet[[names(x = data.return)[i]]] <- CreateAssayObject(counts = data.return[[i]], check.matrix = FALSE)
-          toRet[[names(x = data.return)[i]]] <- CreateAssayObject(counts = data.return[[i]])
+          toRet[[names(x = data.return)[i]]] <- CreateAssayObject(counts = data.return[[i]], check.matrix = FALSE)
           toRet <- SetAssayData(
             object = toRet,
             assay = names(x = data.return)[i],
@@ -1625,40 +1607,86 @@ as.data.frame.Matrix <- function(
 # Internal
 #%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
-# Set a default value if an object is null
-#
-# @param lhs An object to set if it's null
-# @param rhs The value to provide if x is null
-#
-# @return rhs if lhs is null, else lhs
-#
-# @author Hadley Wickham
-# @references https://adv-r.hadley.nz/functions.html#missing-arguments
-#
-`%||%` <- function(lhs, rhs) {
-  if (!is.null(x = lhs)) {
-    return(lhs)
-  } else {
-    return(rhs)
-  }
+.AsList <- function(x) {
+  x <- as.list(x = x)
+  return(sapply(
+    X = unique(x = names(x = x)),
+    FUN = function(i) {
+      return(unlist(
+        x = x[which(x = names(x = x) == i)],
+        recursive = FALSE,
+        use.names = FALSE
+      ))
+    },
+    simplify = FALSE,
+    USE.NAMES = TRUE
+  ))
 }
 
-# Set a default value if an object is NOT null
-#
-# @param lhs An object to set if it's NOT null
-# @param rhs The value to provide if x is NOT null
-#
-# @return lhs if lhs is null, else rhs
-#
-# @author Hadley Wickham
-# @references https://adv-r.hadley.nz/functions.html#missing-arguments
-#
-`%iff%` <- function(lhs, rhs) {
-  if (!is.null(x = lhs)) {
-    return(rhs)
-  } else {
-    return(lhs)
+#' @importFrom ggplot2 cut_number
+#'
+.Cut <- function(min, max, n) {
+  breaks <- levels(x = cut_number(x = c(min, max), n = n))
+  breaks <- gsub(pattern = '.*,', replacement = '', x = breaks)
+  breaks <- gsub(pattern = ']$', replacement = '', x = breaks)
+  as.numeric(x = breaks)
+}
+
+.FindE <- function(x) {
+  x <- as.character(x = x)
+  if (grepl(pattern = 'e', x = x)) {
+    return(as.integer(x = gsub(pattern = '.*e', replacement = '', x = x)))
+  } else if (grepl(pattern = '^0\\.', x = x)) {
+    x <- unlist(x = strsplit(
+      x = gsub(pattern = '.*\\.', replacement = '', x = x),
+      split = ''
+    ))
+    idx <- which(x = x != '0')
+    return(-idx)
   }
+  stop("Invalid format")
+}
+
+#' @importFrom SeuratObject Boundaries
+#'
+.BoundariesByImage <- function(object, fov, boundaries) {
+  if (!is.list(x = boundaries)) {
+    if (is.null(x = names(x = boundaries))) {
+      boundaries <- rep_len(x = list(boundaries), length.out = length(x = fov))
+      names(x = boundaries) <- fov
+    } else {
+      boundaries <- .AsList(x = boundaries)
+    }
+  }
+  if (any(!nchar(x = names(x = boundaries)))) {
+    missing <- setdiff(x = fov, y = names(x = boundaries))
+    idx <- which(x = !nchar(x = names(x = boundaries)))
+    boundaries <- c(
+      boundaries[intersect(x = names(x = boundaries), y = fov)],
+      rep_len(x = boundaries[idx], length.out = length(x = missing))
+    )
+    names(x = boundaries)[!nchar(x = names(x = boundaries))] <- missing
+  }
+  if (any(!fov %in% names(x = boundaries))) {
+    for (i in setdiff(x = fov, y = names(x = boundaries))) {
+      boundaries[[i]] <- Boundaries(object = object[[i]])[1L]
+    }
+  }
+  fov <- union(x = fov, y = names(x = boundaries))
+  if (length(x = boundaries) != length(x = fov)) {
+    fov <- intersect(x = fov, y = names(x = boundaries))
+  }
+  boundaries <- boundaries[fov]
+  for (i in fov) {
+    boundaries[[i]] <- Filter(
+      f = function(x) {
+        return(x %in% Boundaries(object = object[[i]]) || is_na(x = x))
+      },
+      x = boundaries[[i]]
+    )
+  }
+  boundaries <- Filter(f = length, x = boundaries)
+  return(boundaries)
 }
 
 # Generate chunk points
@@ -1894,16 +1922,12 @@ CreateDummyAssay <- function(assay) {
     j = {},
     dims = c(nrow(x = assay), ncol(x = assay))
   )
-  cm <- as(object = cm, Class = "dgCMatrix")
+  cm <- as.sparse(x = cm)
   rownames(x = cm) <- rownames(x = assay)
   colnames(x = cm) <- colnames(x = assay)
-  # TODO: restore once check.matrix is in SeuratObject
-  # return(CreateAssayObject(
-  #   counts = cm,
-  #   check.matrix = FALSE
-  # ))
   return(CreateAssayObject(
-    counts = cm
+    counts = cm,
+    check.matrix = FALSE
   ))
 }
 
@@ -2336,7 +2360,7 @@ RowSumSparse <- function(mat) {
   names(x = output) <- rownames(x = mat)
   return(output)
 }
- 
+
 # Calculate row variance of a sparse matrix
 #
 # @param mat sparse matrix
@@ -2362,7 +2386,7 @@ RowVarSparse <- function(mat) {
 RowSparseCheck <- function(mat) {
   if (!inherits(x = mat, what = "sparseMatrix")) {
     stop("Input should be sparse matrix")
-  } else if (class(x = mat) != "dgCMatrix") {
+  } else if (!is(object = mat, class2 = "dgCMatrix")) {
     warning("Input matrix is converted to dgCMatrix.")
     mat <- as.sparse(x = mat)
   }